+++ NEWS.md
@@ -2,7 +2,10 @@
 
 # development version
 
-- Add Khaowroongrueng 2024 sufentanil ([doi:10.1002/psp4.13205](https://doi.org/10.1002/psp4.13205)) - adult Korean patients undergoing cardiopulmonary bypass surgery.
+- Add Yang 2024 naloxone auto-injector and its four mechanistic opioid-induced
+  respiratory depression reversal models -- buprenorphine, morphine, fentanyl
+  and carfentanil ([doi:10.1002/psp4.13215](https://doi.org/10.1002/psp4.13215))
+  - healthy adults.
 
 - Add Ai 2024 ractopamine ([doi:10.3389/fvets.2024.1399043](https://doi.org/10.3389/fvets.2024.1399043)) - whole-body PBPK in Liaoning cashmere goats.